Env variables

+++ src/main/kotlin/searchengine/searchIndexer/ElasticIndexer.kt
@@ -8,24 +8,30 @@ import libraries.Elastic
 import libraries.PageRepository
 import searchengine.pagerank.Pagerank
 import java.util.concurrent.atomic.AtomicInteger
+import kotlin.math.roundToInt
 
 class ElasticIndexer(
     private val pagerank: Pagerank, private val repository: PageRepository.Client, private val elastic: Elastic
 ) {
-    suspend fun index() = coroutineScope {
-        try {
-            elastic.deleteIndex()
-        } catch (e: Exception) {
-            println("Index not found")
+    suspend fun index(deleteIndex: Boolean) = coroutineScope {
+        if (deleteIndex) {
+            try {
+                println("Checking for existing index")
+                elastic.deleteIndex()
+                println("Deleted previous index")
+            } catch (e: Exception) {
+                println("Index not found")
+            }
+            println("Creating new index")
+            elastic.putMapping()
         }
-        elastic.putMapping()
 
         val count = AtomicInteger(0)
         val totalDocsCount = pagerank.getAll().size
 
 
         val flow = forEachPagerankPage(pagerank.getAll(), repository)
-        for (i in 1..(Runtime.getRuntime().availableProcessors())) {
+        for (i in 1..((Runtime.getRuntime().availableProcessors()).toDouble() / 1.2).roundToInt().coerceAtLeast(1)) {
             launch(Dispatchers.Unconfined) {
                 flow.consumeEach {
 //                    println("$i Indexing ${count.incrementAndGet()}/$totalDocsCount: ${it.second.url}")
@@ -50,7 +56,7 @@ class ElasticIndexer(
 fun CoroutineScope.forEachPagerankPage(
     pagerank: Array<Pagerank.PagerankPage>, repository: PageRepository.Client
 ): ReceiveChannel<Pair<PageRepository.Page?, Pagerank.PagerankPage>> =
-    produce(capacity = (Runtime.getRuntime().availableProcessors()) * 2) {
+    produce(capacity = (Runtime.getRuntime().availableProcessors())) {
         pagerank.forEach {
             if (it.doesForward) {
 //            println("forwards ${it.url}")
